Implemented package structure to avoid common circular import problems and better organize the application.
