+++ LabManager/frequency_manager.py
