+++ LabManager/templates/frequency.html
