+++ LabManager/templates/notices.html
